pyproject.toml
[tool.poetry]
name = "coursework3"
version = "0.1.0"
description = ""
authors = ["Victoria_Zhuravleva <new.juravlik1334@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
pandas = "^2.2.2"
logger = "^1.4"
datetime = "^5.5"
pytest = "^8.2.2"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]

pytest = "^8.2.1"

pytest-cov = "^5.0.0"


[tool.isort]
line_length = 119

[tool.black]
line-length = 119
exclude = ".git"

[tool.mypy]
disallow_untyped_defs=true
implicit_optional=true
warn_return_any=true
exclude = "venv"


[tool.poetry.group.lint.dependencies]
mypy = "^1.10.1"
black = "^24.4.2"
isort = "^5.13.2"
flake8 = "^7.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
